Create strong_password_generator.py
The components of the password are represented in the form of arrays.
Use the random method to select at least one character from each array of characters.
Since the 12-character password is required, so fill the rest of the length of the password left with randomly selected characters from an array formed from the combination of all the characters needed in the final password. Anytime the password is generated, shuffle the password using random.shuffle() to ensure that the final password does not follow a particular pattern.